Move datafilespage to shared space.
