+++ apps/launcher/graphicspage.cpp
@@ -8,8 +8,7 @@
 #include <components/files/configurationmanager.hpp>
 #include <components/files/ogreplugin.hpp>
 #include <components/settings/settings.hpp>
-
-#include "utils/naturalsort.hpp"
+#include <components/file_order_list/utils/naturalsort.hpp>
 
 #include "graphicspage.hpp"
