+++ cmake/OpenMWMacros.cmake
@@ -23,6 +23,22 @@ endforeach (u)
 source_group ("components\\${dir}" FILES ${files})
 endmacro (add_component_dir)
 
+macro (add_component_qt_dir dir)
+set (files)
+foreach (u ${ARGN})
+file (GLOB ALL ${CMAKE_CURRENT_SOURCE_DIR} "${dir}/${u}.[ch]pp")
+foreach (f ${ALL})
+list (APPEND files "${f}")
+list (APPEND COMPONENT_FILES "${f}")
+endforeach (f)
+file (GLOB MOC_H ${CMAKE_CURRENT_SOURCE_DIR} "${dir}/${u}.hpp")
+foreach (fi ${MOC_H})
+list (APPEND COMPONENT_MOC_FILES "${fi}")
+endforeach (fi)
+endforeach (u)
+source_group ("components\\${dir}" FILES ${files})
+endmacro (add_component_qt_dir)
+
 macro (copy_all_files source_dir destination_dir files)
 foreach (f ${files})
 get_filename_component(filename ${f} NAME)
